179,sign fanzhuo,nx18afcl_37lf

+++ shell_test.sh
@@ -200,6 +200,8 @@ function get_images_to_sign_for_MTK()
     echo "boot.img	logo.bin	   system.img
 cache.img	recovery.img  trustzone.bin
 lk.bin	secro.img	   userdata.img
+preloader_nx18afcl_37lf.bin
+preloader_bv303b81t_gb.bin
 "
 }
 
@@ -3039,7 +3041,7 @@ openssl genrsa -out ${keypath}/attest.key -3 2048
     else
             cp $image_unsigned_1_and_2_and_3 $image_signed_new
             echo "COMBINE FINISH******************"
-            echo "     :create $image_signed_new from $image_signed & $image_signed_1 & $image_signed_1"        
+            echo "     :create $image_signed_new from $image_unsigned & $image_signed_1 & $image_signed_1"        
     fi
     #clean
     rm -rf $image_unsigned_1_and_2 $image_unsigned_1_and_2_and_3
@@ -3121,15 +3123,50 @@ openssl genrsa -out ${keypath}/attest.key -3 2048
         image_signed_new="$signed_new_dir""/""$file"
         image_signed_part1="$diff_folder""/""$file"".part1"
         image_signed_part3="$diff_folder""/""$file"".part3"
+        ls $image_signed $image_unsigned $image_signed_new $image_signed_part1 $image_signed_part3
+        if [ ! -f $image_unsigned ] ; then
+            echo "ignore $image_unsigned"
+            continue
+        fi
         # 0x4040 = 16448
         size_image_signed_part1=16448
-        $script 159 $image_signed $image_unsigned $image_signed_new $size_image_signed_part1 $image_signed_part1 $image_signed_part3 $check $combine
-        if [ $? -ne 0 ] ; then
-            # clean
-            if [ "$combine" != "true" ] ; then
-                rm -rf "$signed_new_dir"
+        echo $image_signed | grep "preloader_".*".bin"
+        if [ $? -eq 0 ] ; then
+            echo "special handler for preloader.bin"
+            if [ "$check" != "true" ] ; then
+                echo "SUCCESS******************"
+                echo "     :create $image_signed_part1 from $image_signed"
+                echo cp $image_signed $image_signed_part1
+                cp $image_signed $image_signed_part1
+                echo cp $image_signed $image_signed_part3
+                cp $image_signed $image_signed_part3
+            else
+                if [ "$combine" != "true" ] ; then
+                    diff $image_signed_part1 $image_signed
+                    if [ $? -ne 0 ] ; then
+                        echo "ERROR******************"
+                        echo "     :$image_signed_part1 $image_signed are different"
+                        exit 1
+                    else
+                        echo "SUCCESS******************"
+                        echo "     :create $image_signed_new from $image_signed & $image_unsigned"
+                    fi
+                else
+                    echo "COMBINE FINISH******************"
+                    echo "     :create $image_signed_new from $image_signed_part1"        
+                    echo cp $image_signed_part1 $image_signed_new
+                    cp $image_signed_part1 $image_signed_new
+                fi
             fi
+        else
+            $script 159 $image_signed $image_unsigned $image_signed_new $size_image_signed_part1 $image_signed_part1 $image_signed_part3 $check $combine
+            if [ $? -ne 0 ] ; then
+                # clean
+                if [ "$combine" != "true" ] ; then
+                    rm -rf "$signed_new_dir"
+                fi
             exit 1;
+            fi
         fi
     done
     # clean
@@ -3243,7 +3280,7 @@ openssl genrsa -out ${keypath}/attest.key -3 2048
     else
             cp $image_unsigned_1_and_2_and_3 $image_signed_new
             echo "COMBINE FINISH******************"
-            echo "     :create $image_signed_new from $image_signed & $image_signed_1 & $image_signed_1"        
+            echo "     :create $image_signed_new from $image_unsigned & $image_signed_1 & $image_signed_1"        
     fi
     #clean
     rm -rf $image_unsigned_1_and_2 $image_unsigned_1_and_2_and_3
@@ -3596,7 +3633,7 @@ openssl genrsa -out ${keypath}/attest.key -3 2048
     echo "cd MTK/mtk_release/sign_image_split/sign-image"
     pwd
     sed -i s/"^MTK_PROJECT_NAME := .*"/"MTK_PROJECT_NAME := $project"/g Android.mk
-    sed -i s/"^MTK_PROJECT_NAME := .*"/"MTK_PROJECT_NAME := $project"/g Android.mk
+    sed -i s/"^MTK_PROJECT := .*"/"MTK_PROJECT := $project"/g Android.mk
     make -f Android.mk
     cd $current_dir
 
@@ -3880,6 +3917,127 @@ openssl genrsa -out ${keypath}/attest.key -3 2048
     else
         echo "e : $? err: $err"
     fi    
+    ;;
+179)   echo "given folder1 unsigned & project_name"
+    echo "sign unsigned.zip"
+    echo "get diff between unsigned.zip & signed.zip"
+    echo "create filelist.txt"
+    echo "MTK, fanzhuo"
+    echo "/mnt/sign_server/sign_bv303b81t_gb.sh"
+    echo "similar with 163"
+    echo "run on /mnt/sign_server"
+    unsigned_zip=$2
+    project=$3
+    current_dir=`pwd`
+    sign_workspace="/tmp/"
+    
+    zip_folder=`dirname $unsigned_zip`
+    zip_name=`basename $unsigned_zip`
+    postfix=`echo $zip_name|awk -F '.' '{print $NF}'`
+    zip_filename=`echo $zip_name|awk -F "[.]$postfix" '{print $1}'`
+    #
+#   signed_zip="$sign_workspace""/""$zip_filename""_signed"".""$postfix"
+#   diff_zip="$sign_workspace""/""$zip_filename""_diff"".""$postfix"
+    signed_zip="$4"
+    diff_zip="$5"
+    signed_list_file="$6"
+    signed_md5_file="$7"
+    # clean & mkdir
+    unsigned_unzip="$current_dir""/MTK/mtk_release/sign_image_split/out/target/product/""$project"
+    diff_folder="$unsigned_unzip""/diff/"
+    signed_new_folder="$unsigned_unzip""/signed_new/"
+    signed_bin_folder="$unsigned_unzip""/signed_bin"
+    signed_bin_tmp_folder="$unsigned_unzip""/signed_bin_tmp"
+    verity_folder="$unsigned_unzip""/verity"
+    rm -rf $signed_list_file
+    rm -rf $signed_md5_file
+    rm -rf $unsigned_unzip
+    rm -rf $diff_folder
+    rm -rf $signed_new_folder
+    rm -rf $signed_bin_tmp_folder
+    rm -rf $verity_folder
+    mkdir $unsigned_unzip
+    mkdir -p $diff_folder
+    mkdir -p $signed_new_folder
+    mkdir -p $verity_folder
+
+    # unzip
+    unzip $unsigned_zip -d $unsigned_unzip
+    ls -l $unsigned_unzip
+    # sign
+    cd $current_dir"/MTK/mtk_release/sign_image_split/sign-image"
+    echo "cd MTK/mtk_release/sign_image_split/sign-image"
+    pwd
+    sed -i s/"^MTK_PROJECT_NAME := .*"/"MTK_PROJECT_NAME := $project"/g Android.mk
+    sed -i s/"^MTK_PROJECT := .*"/"MTK_PROJECT := $project"/g Android.mk
+    make -f Android.mk
+    cd $current_dir
+
+    cd $current_dir"/MTK/mtk_release/PreloaderResign/"
+    echo ./build.sh $unsigned_unzip"/preloader_""$project"".bin" $signed_bin_folder
+    ./build.sh $unsigned_unzip"/preloader_""$project"".bin" $signed_bin_folder
+    cd $current_dir
+
+    # store signed_bin/
+    cp -r $signed_bin_folder $signed_bin_tmp_folder 
+
+    # handle with signed images
+    # mv xxx-sign.xxx xxx.xxx
+    for file in `dir $signed_bin_tmp_folder`
+    do
+        filename=`echo $file|awk -F "-sign" '{print $1$2}'`;
+        file_unsigned="$unsigned_unzip""/""$filename"
+        file_signed="$signed_bin_tmp_folder""/""$file"
+        echo $filename >> $signed_list_file
+        mv $file_signed "$signed_bin_tmp_folder""/""$filename"
+    done
+
+    # make diff
+    $script 160 "$signed_bin_tmp_folder" $unsigned_unzip $diff_folder $signed_new_folder
+    if [ $? -ne 0 ] ; then
+#        rm -rf $unsigned_unzip
+#        rm -rf $diff_folder
+#        rm -rf $signed_new_folder
+        exit 1;
+    fi
+    
+    cd $diff_folder
+    zip $diff_zip ./*
+    cd $current_dir
+
+    # override unsigned images with signed ones
+#    cp "$signed_bin_tmp_folder""/"* "$unsigned_unzip""/"
+
+    cp $unsigned_unzip"/"*"-verified"* $verity_folder"/"
+
+    # clean
+#    rm -rf $unsigned_unzip"/sig/"
+#    rm -rf $signed_bin_tmp_folder
+#    rm $unsigned_unzip"/"*"-verified"*
+#    rm -rf $diff_folder
+#    rm -rf $signed_new_folder
+
+#    cd $unsigned_unzip
+    cd $signed_bin_folder
+    for f in `ls`
+    do
+        md5sum $f
+        md5=`md5sum $f |awk '{print $1}'` 
+        echo "$md5""   ""$f" |tee -a $signed_md5_file
+    done
+#    zip $signed_zip ./*
+
+    cd $unsigned_unzip
+    echo zip -r $signed_zip signed_bin/
+    zip -r $signed_zip signed_bin/
+    echo zip -r $signed_zip verity/
+    zip -r $signed_zip verity/
+    echo zip -r $signed_zip "sig/"
+    zip -r $signed_zip "sig/"
+    cd $current_dir
+
+    # clean
+#    rm -rf $unsigned_unzip
     ;;
 *) echo "others"
     echo "1: $2"
